requirements.txt
pymongo[srv]
python-dotenv
Flask
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
APScheduler
mistralai
flask-cors
openai